wip: move recipe to clean arch
