+++ lib/features/recipe/business/entities/recipe_entity.dart
@@ -1,8 +1,13 @@
+import 'package:cooking_friend/features/recipe/data/models/recipe_model.dart';
+
 class RecipeEntity {
-  final int id;
+  final int? id;
   final String? name;
   final bool? isFavorite;
 
-  RecipeEntity(
-      this.id, this.name, this.isFavorite);
+  RecipeEntity(this.id, this.name, this.isFavorite);
+
+  RecipeModel toModel([int? newId]) {
+    return RecipeModel(newId ?? id, name, isFavorite);
+  }
 }
