+++ lib/features/recipe/business/entities/recipe_step_entity.dart
@@ -1,6 +1,3 @@
-import 'package:cooking_friend/constants.dart';
-import 'package:cooking_friend/features/recipe/data/models/recipe_model.dart';
-
 class RecipeStepEntity {
   final int id;
   final String? step;
