+++ lib/features/recipe/business/usecases/recipe_use_case.dart
@@ -1,8 +1,6 @@
-import 'package:cooking_friend/constants.dart';
+import 'package:cooking_friend/skeleton/constants.dart';
 import 'package:cooking_friend/features/recipe/business/entities/recipe_entity.dart';
-import 'package:cooking_friend/features/recipe/business/entities/recipe_ingredient_entity.dart';
 import 'package:cooking_friend/features/recipe/business/entities/recipe_modification_entity.dart';
-import 'package:cooking_friend/features/recipe/business/entities/recipe_step_entity.dart';
 import 'package:cooking_friend/features/recipe/data/repositories/i_recipe_repository_implementation.dart';
 import 'package:cooking_friend/features/recipe/presentation/provider/recipe_getx.dart';
 import 'package:flutter/material.dart';
@@ -38,8 +36,11 @@ class RecipeUseCase {
     }
   }
 
-  Future<bool> save(GlobalKey<FormBuilderState> formKey, BuildContext context,
-      List<RecipeModificationEntity> lstRecipeModification, bool isFavorite) async {
+  Future<bool> save(
+      GlobalKey<FormBuilderState> formKey,
+      BuildContext context,
+      List<RecipeModificationEntity> lstRecipeModification,
+      bool isFavorite) async {
     return await _save(formKey, context, lstRecipeModification, isFavorite)
         .then((success) {
       if (success) {
@@ -55,10 +56,11 @@ class RecipeUseCase {
     await updateList("/recipeManagement", context);
   }
 
-  Future<void> _delete(List<RecipeModificationEntity> lstStorageItemModification,
+  Future<void> _delete(
+      List<RecipeModificationEntity> lstStorageItemModification,
       BuildContext context) async {
     await recipeRepository.deleteRecipe(id: recipeGetx.currentId).then(
-          (res) {
+      (res) {
         lstStorageItemModification.add(RecipeModificationEntity()
           ..id = recipeGetx.currentId
           ..action = RecipeManagementAction.delete
@@ -74,36 +76,38 @@ class RecipeUseCase {
     if (recipeGetx.action == RecipeManagementAction.edit.name.obs) {
       await recipeRepository
           .updateRecipe(
-          recipe,
-          recipeGetx.currentId,
-          recipeGetx.ingredientsToRemove,
-          recipeGetx.stepsToRemove)
+        recipe: recipe.toModel(),
+      )
           .then((res) {
         lstRecipeModification.add(RecipeModificationEntity()
           ..id = recipeGetx.currentId
           ..action = RecipeManagementAction.edit
-          ..item = recipe);
+          ..item = recipe.toModel());
       });
     } else {
-      await recipeRepository.saveNewRecipe(recipe).then((res) {
-        recipe.id = res;
-        lstRecipeModification.add(RecipeModification()
-          ..id = res
-          ..action = RecipeManagementAction.add
-          ..item = recipe);
+      await recipeRepository
+          .saveNewRecipe(recipe: recipe.toModel())
+          .then((either) {
+        either.fold((failure) {}, (newId) {
+          lstRecipeModification.add(RecipeModificationEntity()
+            ..id = newId
+            ..action = RecipeManagementAction.add
+            ..item = recipe.toModel(newId));
+        });
       });
     }
     formKey.currentState!.reset();
   }
 
-  Future<bool> _save(GlobalKey<FormBuilderState> formKey, BuildContext context,
-      List<RecipeModificationEntity> lstRecipeModification, bool isFavorite) async {
+  Future<bool> _save(
+      GlobalKey<FormBuilderState> formKey,
+      BuildContext context,
+      List<RecipeModificationEntity> lstRecipeModification,
+      bool isFavorite) async {
     // ne pas saver ce qui ont le meme id
     if (formKey.currentState!.saveAndValidate()) {
-      RecipeEntity newRecipe = RecipeEntity()
-        ..name = formKey.currentState?.value["recipe_title"]
-        ..isFavorite = isFavorite;
-      for (int i = 0; i < recipeGetx.steps.length; i++) {
+      RecipeEntity newRecipe = RecipeEntity(null, formKey.currentState?.value["recipe_title"], isFavorite);
+      /*for (int i = 0; i < recipeGetx.steps.length; i++) {
         var currentElement = recipeGetx.steps[i];
         var content = formKey.currentState?.value["rs_${currentElement.guid}"];
         RecipeStepEntity step = RecipeStepEntity()..step = content;
@@ -113,23 +117,23 @@ class RecipeUseCase {
         }
 
         newRecipe.lstSteps.add(step);
-      }
+      }*/
 
-      for (int i = 0; i < recipeGetx.ingredients.length; i++) {
+      /*for (int i = 0; i < recipeGetx.ingredients.length; i++) {
         var currentElement = recipeGetx.ingredients[i];
         RecipeIngredientEntity ri = RecipeIngredientEntity();
         ri.ingredient =
-        formKey.currentState?.value["ri_${currentElement.guid}"];
+            formKey.currentState?.value["ri_${currentElement.guid}"];
         ri.measuringUnit =
-        formKey.currentState?.value["riu_${currentElement.guid}"];
-        ri.quantity = double
-            .parse(formKey.currentState?.value["riq_${currentElement.guid}"]);
+            formKey.currentState?.value["riu_${currentElement.guid}"];
+        ri.quantity = double.parse(
+            formKey.currentState?.value["riq_${currentElement.guid}"]);
         if (currentElement.ingredient != null) {
           ri.id = currentElement.ingredient!.id;
         }
         ri.order = i;
         newRecipe.lstIngredients.add(ri);
-      }
+      }*/
 
       await _saveAndUpdate(newRecipe, lstRecipeModification, formKey)
           .then((res) {
