+++ lib/features/storage/data/models/storage_modification.dart
@@ -1,4 +1,4 @@
-import 'package:cooking_friend/constants.dart';
+import 'package:cooking_friend/skeleton/constants.dart';
 import 'package:cooking_friend/features/storage/data/models/storage_model.dart';
 
 class StorageItemModification {
