--- lib/getx/services/recipe_service.dart
@@ -1,19 +1,0 @@
-import 'package:cooking_friend/constants.dart';
-import 'package:cooking_friend/features/recipe/presentation/provider/recipe_getx.dart';
-import 'package:cooking_friend/features/recipe/data/models/recipe_model.dart';
-import 'package:cooking_friend/features/recipe/data/models/recipe_ingredient_model.dart';
-import 'package:cooking_friend/features/recipe/business/entities/recipe_modification_entity.dart';
-import 'package:cooking_friend/features/recipe/data/models/recipe_step_model.dart';
-import 'package:cooking_friend/getx/services/isar_service.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:get/get.dart';
-
-class RecipeService {
-
-  //#region Private
-
-  
-
-//#endregion
-}
