+++ lib/global_widget/barcode_scanner.dart
