+++ lib/global_widget/search_display_card.dart
