--- lib/screens/shoppingList/shopping_list.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ShoppingList extends StatefulWidget {
-  const ShoppingList({super.key});
-
-  @override
-  State<ShoppingList> createState() => _ShoppingListState();
-}
-
-class _ShoppingListState extends State<ShoppingList> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).appBarTheme.backgroundColor,
-        title: const Text(
-          "Shopping list",
-        ),
-      ),
-      body: const Column(children: [],),
-    );
-  }
-}
