+++ lib/skeleton/constants.dart
