+++ lib/skeleton/main_wrapper.dart
@@ -1,16 +1,16 @@
+import 'package:cooking_friend/features/recipe/data/repositories/i_recipe_repository_implementation.dart';
 import 'package:cooking_friend/features/storage/data/repositories/i_storage_repository_implementation.dart';
-import 'package:cooking_friend/navigation/recipe_navigation.dart';
+import 'package:cooking_friend/skeleton/navigation/recipe_navigation.dart';
 import 'package:cooking_friend/skeleton/navigation/storage_navigation.dart';
-import 'package:cooking_friend/getx/services/isar_service.dart';
-import 'package:cooking_friend/theme/custom_theme.dart';
+import 'package:cooking_friend/skeleton/theme/custom_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 class MainWrapper extends StatefulWidget {
-  final IsarService service;
-  final IStorageRepositoryImplementation repository;
+  final IStorageRepositoryImplementation storageRepository;
+  final IRecipeRepositoryImplementation recipeRepository;
 
-  const MainWrapper(this.service, this.repository, {super.key});
+  const MainWrapper(this.storageRepository, this.recipeRepository, {super.key});
 
   @override
   State<MainWrapper> createState() => _MainWrapperState();
@@ -80,8 +80,8 @@ class _MainWrapperState extends State<MainWrapper> {
           child: IndexedStack(
             index: _selectedIndex,
             children: [
-              StorageNavigation(widget.repository),
-              RecipeNavigation(widget.service),
+              StorageNavigation(widget.storageRepository),
+              RecipeNavigation(widget.recipeRepository),
               //ShoppingListNavigation(widget.service),
             ],
           ),
